refactor: addNotification 함수 hook으로 변환, props drilling 해소
