+++ src/advanced/App.tsx
@@ -1,4 +1,3 @@
-import { useCallback } from "react";
 import { Product } from "../types";
 import { AdminPage } from "./components/AdminPage";
 import { CartPage } from "./components/CartPage";
@@ -6,7 +5,7 @@ import { UIToast } from "./components/ui/UIToast";
 import { Layout } from "./components/layout/Layout";
 import { useDebounce } from "./utils/hooks/useDebounce";
 import { useAtom } from "jotai";
-import { searchTermAtom, isAdminAtom, notificationsAtom } from "./atoms";
+import { searchTermAtom, isAdminAtom } from "./atoms";
 
 export interface ProductWithUI extends Product {
   description?: string;
@@ -22,37 +21,17 @@ export interface Notification {
 const App = () => {
   const [isAdmin, setIsAdmin] = useAtom(isAdminAtom);
   const [searchTerm, setSearchTerm] = useAtom(searchTermAtom);
-  const [notifications, setNotifications] = useAtom(notificationsAtom);
-
   const debouncedSearchTerm = useDebounce(searchTerm, 500);
-
-  const addNotification = useCallback(
-    (message: string, type: "error" | "success" | "warning" = "success") => {
-      const id = Date.now().toString();
-      setNotifications((prev) => [...prev, { id, message, type }]);
-
-      setTimeout(() => {
-        setNotifications((prev) => prev.filter((n) => n.id !== id));
-      }, 3000);
-    },
-    []
-  );
 
   return (
     <Layout>
-      <UIToast
-        notifications={notifications}
-        setNotifications={setNotifications}
-      />
+      <UIToast />
       <Layout.Header />
       <Layout.Main>
         {isAdmin ? (
-          <AdminPage addNotification={addNotification} />
+          <AdminPage />
         ) : (
-          <CartPage
-            addNotification={addNotification}
-            debouncedSearchTerm={debouncedSearchTerm}
-          />
+          <CartPage debouncedSearchTerm={debouncedSearchTerm} />
         )}
       </Layout.Main>
     </Layout>
